Added Css
First of all, I changed the UI to look more stylish with CSS3 and Flex Containers.
Then I did a couple of changes here and there to make it work in node, and I modified the package.json and yarn.lock files to be the same as in the github repo.