Separation of concerns and functional Login
Well, first of all, I'd separate the app in 3 different html's. 

The first on is home (index.html), which is the entry point of the app, it just has the header and the text that the last version has and a navbar with for navigation trough pages. 
Sign in, it's the same form that we had before but now it has it's own new and fancy page. 
And for last but not least, the login page, where you put the user's email and password. Then Node asks MongoDB for the information of that email account, it  hash and salt the input password using the same salt of the saved password, compares the saved password and the inputs password. And if it's all correct, the user is sent back to index and on the console you'll see the message "LogIn made successfully!".